backend/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
openai
tiktoken
python-dotenv
rich
faiss-cpu
pypdf
markdown
